Refactoring
- moved Delegates to ITransitionProvider
