+++ ChitChat02/PresentationLayer/Animations+Transitions/ViewController Transitions/TransitionProvider.swift
@@ -9,16 +9,27 @@
 import Foundation
 import UIKit
 
-protocol IViewControllerTransition: UIViewControllerAnimatedTransitioning {
-    var presenting: Bool { get set }
-}
+class TransitionProvider: NSObject, ITransitionProvider {
+    private lazy var presentTransition: IViewControllerTransition = TopLeftRightBottom()
+    private lazy var pushTransition: IViewControllerTransition = RotatingCard()
 
-protocol ITransitionProvider {
-    var presentTransition: IViewControllerTransition { get }
-    var pushTransition: IViewControllerTransition { get }
-}
-
-class TransitionProvider: ITransitionProvider {
-    lazy var presentTransition: IViewControllerTransition = TopLeftRightBottom()
-    lazy var pushTransition: IViewControllerTransition = RotatingCard()
+    func navigationController(_ navigationController: UINavigationController,
+                              animationControllerFor operation: UINavigationController.Operation,
+                              from fromVC: UIViewController,
+                              to toVC: UIViewController) -> UIViewControllerAnimatedTransitioning? {
+        pushTransition.presenting = (operation == .push)
+        return pushTransition
+    }
+    
+    func animationController(forPresented presented: UIViewController,
+                             presenting: UIViewController,
+                             source: UIViewController) -> UIViewControllerAnimatedTransitioning? {
+        presentTransition.presenting = true
+        return presentTransition
+    }
+    
+    func animationController(forDismissed dismissed: UIViewController) -> UIViewControllerAnimatedTransitioning? {
+        presentTransition.presenting = false
+        return presentTransition
+    }
 }
